bổ sung các trang footer (trừ others), tổ chức lại code trong layout (chia rõ ràng footers và header), sửa chính tả experience
