+++ Frontend/src/pages/Experience/Experience.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useCallback } from 'react';
-import { useUrlFilters } from '../hooks/useUrlFilters';
-import { fetchExperiences } from '../services/searchApi';
-import type { ExperienceResult } from '../types/search';
+import { useUrlFilters } from '../../hooks/useUrlFilters';
+import { fetchExperiences } from '../../services/searchApi';
+import type { ExperienceResult } from '../../types/search';
 
 const fmt = (price: number) =>
     new Intl.NumberFormat('vi-VN', { style: 'currency', currency: 'VND' }).format(price);
