+++ Frontend/src/pages/Transport/Trains.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useCallback } from 'react';
-import { useUrlFilters } from '../hooks/useUrlFilters';
-import { fetchTrains } from '../services/searchApi';
-import type { TrainResult } from '../types/search';
+import { useUrlFilters } from '../../hooks/useUrlFilters';
+import { fetchTrains } from '../../services/searchApi';
+import type { TrainResult } from '../../types/search';
 
 const fmt = (price: number) =>
     new Intl.NumberFormat('vi-VN', { style: 'currency', currency: 'VND' }).format(price);
